requirements.txt
Package                       Version
----------------------------- -------
asgiref                       3.8.1
Django                        5.0.6
django-filter                 24.2
djangorestframework           3.15.2
djangorestframework-simplejwt 5.3.1
drf-yasg                      1.21.7
inflection                    0.5.1
packaging                     24.1
pip                           24.1.1
PyJWT                         2.8.0
pytz                          2024.1
PyYAML                        6.0.1
setuptools                    70.2.0
sqlparse                      0.5.0
uritemplate                   4.1.1
